refactor: 카드 생성기 인터페이스로 추상화
