+++ src/main/java/blackjack/domain/card/generator/CardDeckGenerator.java
@@ -1,0 +1,9 @@
+package blackjack.domain.card.generator;
+
+import blackjack.domain.card.Card;
+import java.util.List;
+
+public interface CardDeckGenerator {
+
+    List<Card> generate();
+}
