+++ src/main/java/blackjack/domain/card/generator/RandomCardDeckGenerator.java
@@ -1,10 +1,13 @@
-package blackjack.domain.card;
+package blackjack.domain.card.generator;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardPattern;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 
-public class CardDeckGenerator {
+public class RandomCardDeckGenerator implements CardDeckGenerator{
 
     private static final List<Card> CARDS = initializeCardDeckByCardNumber();
 
@@ -22,8 +25,9 @@ public class CardDeckGenerator {
         }
     }
 
-    public LinkedList<Card> generate() {
-        LinkedList<Card> cards = new LinkedList<>(CARDS);
+    @Override
+    public List<Card> generate() {
+        List<Card> cards = new LinkedList<>(CARDS);
         Collections.shuffle(cards);
         return cards;
     }
