+++ src/test/java/blackjack/domain/card/CardDeckTest.java
@@ -3,6 +3,7 @@ package blackjack.domain.card;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.domain.card.generator.RandomCardDeckGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,7 +12,7 @@ public class CardDeckTest {
     @Test
     @DisplayName("사용한 카드는 제거한다")
     void drawCard() {
-        final CardDeck cardDeck = new CardDeck(new CardDeckGenerator());
+        final CardDeck cardDeck = new CardDeck(new RandomCardDeckGenerator());
         final int deckSize = 52;
 
         for (int i = 0; i < deckSize; i++) {
@@ -26,7 +27,7 @@ public class CardDeckTest {
     @Test
     @DisplayName("처음 플레이어에게 draw 하는 카드 수는 2장이다.")
     void drawInitialCards() {
-        final CardDeck cardDeck = new CardDeck(new CardDeckGenerator());
+        final CardDeck cardDeck = new CardDeck(new RandomCardDeckGenerator());
         final int expected = 2;
 
         final Cards cards = cardDeck.drawInitialCards();
